+++ tests/Jellyfin.Plugin.Ingest.Tests/QuarantineListingTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using System.Linq;
+using Jellyfin.Plugin.Ingest.Quarantine;
 using Jellyfin.Plugin.Ingest.Service;
 using Xunit;